types changes

+++ libs/openFrameworks/types/ofBaseTypes.cpp
@@ -335,7 +335,7 @@ bool ofBaseMaterial::isBound() const {
 
 
 void ofBaseSoundInput::audioIn( ofSoundBuffer& buffer ){
-	audioIn(&buffer[0], buffer.getNumFrames(), buffer.getNumChannels(), buffer.getDeviceID(), buffer.getTickCount());
+	audioIn(&buffer[0], static_cast<int>(buffer.getNumFrames()), static_cast<int>(buffer.getNumChannels()), buffer.getDeviceID(), buffer.getTickCount());
 }
 
 void ofBaseSoundInput::audioIn( float * input, int bufferSize, int nChannels, int deviceID, long unsigned long tickCount ){
@@ -348,7 +348,7 @@ void ofBaseSoundInput::audioIn( float * input, int bufferSize, int nChannels ){
 
 
 void ofBaseSoundOutput::audioOut( ofSoundBuffer& buffer ){
-	audioOut(&buffer[0], buffer.getNumFrames(), buffer.getNumChannels(), buffer.getDeviceID(), buffer.getTickCount());
+	audioOut(&buffer[0], static_cast<int>(buffer.getNumFrames()), static_cast<int>(buffer.getNumChannels()), buffer.getDeviceID(), buffer.getTickCount());
 }
 
 void ofBaseSoundOutput::audioOut( float * output, int bufferSize, int nChannels, int deviceID, long unsigned long tickCount  ){
